requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
clickclick==20.10.2
colorama==0.4.4
connexion==2.7.0
discord.py==1.7.3
Flask==2.0.1
gql==3.0.0a5
graphql-core==3.1.5
idna==2.10
inflection==0.5.1
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
multidict==5.1.0
NamedAtomicLock==1.1.3
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pyrsistent==0.17.3
python-dotenv==0.18.0
PyYAML==5.4.1
reportlab==3.5.67
requests==2.25.1
six==1.16.0
tabulate==0.8.9
typing-extensions==3.10.0.0
urllib3==1.26.5
websockets==8.1
Werkzeug==2.0.1
yarl==1.6.3
